Secure api

+++ apis/posts.ts
@@ -1,17 +1,26 @@
 import express from 'express';
 import { v4 as uuidv4 } from "uuid";
 import getFromDatabase from "../lib/getFromDatabase";
-import {pool} from "../client";
+import { pool } from "../client";
+import authorisation from "../lib/authorisation";
 const router = express.Router();
 
 router.use(express.json());
 
-router.get('/', async (_req, res) => {
-	await getFromDatabase('posts', res)
+router.get('/', async (req, res) => {
+	const token = (req.header('Authorization' as string)?.split(' ')[1])
+
+	if (authorisation(token, res)) {
+		await getFromDatabase('posts', res)
+	}
 })
 
 router.get('/:itemId', async (req, res) => {
 	const itemId = req.params.itemId
+	const token = (req.header('Authorization' as string)?.split(' ')[1])
+
+	if (authorisation(token, res)) {
+
 
 	const client = await pool.connect();
 
@@ -46,6 +55,7 @@ router.get('/:itemId', async (req, res) => {
 	}
 
 	res.status(200).send(postTemplate)
+	}
 })
 
 router.post('/', async (req, res) => {
@@ -53,34 +63,38 @@ router.post('/', async (req, res) => {
 	const id = uuidv4()
 
 	if (req.body) {
+		const token = (req.header('Authorization' as string)?.split(' ')[1])
+
 		const {notes, photo, peakId, authorId} = req.body
 
-		const client = await pool.connect();
+		if (authorisation(token, res, authorId)) {
+			const client = await pool.connect();
 
-		if (client) {
-			const author = await client.query('SELECT id, username, firstname, avatar, suspension_timeout, is_banned, role_id FROM users WHERE id=($1)', [authorId]).then(response => {
-				return response.rows[0]
-			})
+			if (client) {
+				const author = await client.query('SELECT id, username, firstname, avatar, suspension_timeout, is_banned, role_id FROM users WHERE id=($1)', [authorId]).then(response => {
+					return response.rows[0]
+				})
 
-			const peak = await client.query(`SELECT * FROM peaks WHERE id = ($1)`, [peakId]).then(response => {
-				return response.rows[0]
-			})
+				const peak = await client.query(`SELECT * FROM peaks WHERE id = ($1)`, [peakId]).then(response => {
+					return response.rows[0]
+				})
 
-			await client.query(`INSERT INTO posts (id, created_at, notes, photo, peak_id, is_hidden, author_id) VALUES ('${id}', '${now}', '${notes}', '${photo}', '${peak.id}', '${false}', '${author.id}') ON CONFLICT DO NOTHING;`).then(() => {
-				const newPost = {
-					id,
-					createdAt: new Date(now),
-					notes,
-					photo,
-					peak,
-					isHidden: false,
-					author,
-				}
+				await client.query(`INSERT INTO posts (id, created_at, notes, photo, peak_id, is_hidden, author_id) VALUES ('${id}', '${now}', '${notes}', '${photo}', '${peak.id}', '${false}', '${author.id}') ON CONFLICT DO NOTHING;`).then(() => {
+					const newPost = {
+						id,
+						createdAt: new Date(now),
+						notes,
+						photo,
+						peak,
+						isHidden: false,
+						author,
+					}
 
-				res.status(200).send(newPost)
-			})
-		} else {
-			res.status(500).send('Connection failed');
+					res.status(200).send(newPost)
+				})
+			} else {
+				res.status(500).send('Connection failed');
+			}
 		}
 	} else {
 		res.status(400).send('Request failed')
@@ -89,6 +103,8 @@ router.post('/', async (req, res) => {
 
 router.put('/:itemId', async (req, res) => {
 	if (req.params.itemId && req.body) {
+		const token = (req.header('Authorization' as string)?.split(' ')[1])
+
 		const client = await pool.connect()
 
 		if (client) {
@@ -98,30 +114,30 @@ router.put('/:itemId', async (req, res) => {
 			const post = await client.query(`SELECT * FROM posts WHERE id=($1)`, [itemId]).then(response => {
 				return response.rows[0]
 			})
-			const author = await client.query(`SELECT * FROM users WHERE id=($1)`, [post.autor_id]).then(response => {
+			const author = await client.query(`SELECT * FROM users WHERE id=($1)`, [post.author_id]).then(response => {
 				return response.rows[0]
 			})
 			const peak = await client.query(`SELECT * FROM peaks WHERE id=($1)`, [post.peak_id]).then(response => {
 				return response.rows[0]
 			})
 
-			const updatedPost = {
-				id: itemId,
-				createdAt: post.created_at,
-				notes: notes ?? post.notes,
-				photo: photo ?? post.photo,
-				peak,
-				isHidden: isHidden === undefined ? post.is_hidden : isHidden,
-				author,
+			if (authorisation(token, res, author.id)) {
+				const updatedPost = {
+					id: itemId,
+					createdAt: post.created_at,
+					notes: notes ?? post.notes,
+					photo: photo ?? post.photo,
+					peak,
+					isHidden: isHidden === undefined ? post.is_hidden : isHidden,
+					author,
+				}
+
+				await client.query(`UPDATE posts SET notes='${updatedPost.notes}', photo='${updatedPost.photo}', is_hidden='${updatedPost.isHidden}' WHERE id=($1)`, [itemId]).then(() => {
+					res.status(200).send(updatedPost);
+				}).catch((err) => {
+					res.status(500).send(`Connection failed: ${err.message}`);
+				})
 			}
-
-			await client.query(`UPDATE posts SET notes='${updatedPost.notes}', photo='${updatedPost.photo}', is_hidden='${updatedPost.isHidden}' WHERE id=($1)`, [itemId]).then(() => {
-				res.status(200).send(updatedPost);
-			}).catch((err) => {
-				res.status(500).send(`Connection failed: ${err.message}`);
-			})
-
-
 		} else {
 			res.status(500).send('Connection failed');
 		}
@@ -132,18 +148,22 @@ router.put('/:itemId', async (req, res) => {
 
 router.delete('/:itemId', async (req, res) => {
 	if (req.params.itemId) {
-		const client = await pool.connect();
+		const token = (req.header('Authorization' as string)?.split(' ')[1])
 
-		if (client) {
-			const itemId = req.params.itemId
+		if (authorisation(token, res)) {
+			const client = await pool.connect();
 
-			await client.query(`DELETE FROM posts WHERE id=($1)`, [itemId]).then(() => {
-				res.status(200).send('Item deleted');
-			}).catch((err) => {
-				res.status(500).send(`Connection failed: ${err.message}`);
-			})
-		} else {
-			res.status(500).send('Connection failed');
+			if (client) {
+				const itemId = req.params.itemId
+
+				await client.query(`DELETE FROM posts WHERE id=($1)`, [itemId]).then(() => {
+					res.status(200).send('Item deleted');
+				}).catch((err) => {
+					res.status(500).send(`Connection failed: ${err.message}`);
+				})
+			} else {
+				res.status(500).send('Connection failed');
+			}
 		}
 	} else {
 		res.status(400).send('Request failed');
